Tried making gtest work with template types
Fails, because gtest uses typedef all over the place, which at some point breaks template compatibility. I guess already when defining MyTypes it fails, since there the template arguments of the "_front" classes are already lost. If that were done using an alias and the aliasing was propagated properly, it might work, but that would be a gtest overhaul.